pyproject.toml
[tool.poetry]
name = "django-materialized-view"
version = "0.1.0"
description = ""
authors = ["Farruh Sheripov <sheripov.farruh@gmail.com>"]
readme = "README.md"
packages = [{include = "django_materialized_view"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"
pytest-subtests = "^0.10.0"
factory-boy = "^3.2.1"
pre-commit = "^3.1.1"
pytest = "^7.2.2"
psycopg2-binary = "^2.9.5"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "testproject.settings"
testpaths = "tests"
